Move AddressManager to an internal package
